fix multiple message handling problem
We need to re-subsribe to topics each time the mqtt connection
reconnects, however we had also move the subscription to the
mqtt on message event inside the connect event handler.  The result
was that we would end up processing each message once for each
time there had been a disonnect/reconnect to the mqtt server.
A bit difficult to diagnose since you'd not see the problem
when you stopped/started the micro-app to debug but there
were finally enough clues in the observed behaviour to
figure this out. Move the registration of these messages
outside the connect function.